move project-related classes to their own package
