+++ src/main/groovy/com/blackbaud/templates/project/BuildFile.groovy
@@ -1,4 +1,4 @@
-package com.blackbaud.templates
+package com.blackbaud.templates.project
 
 
 class BuildFile extends ProjectFile {
