+++ src/main/groovy/com/blackbaud/templates/project/PerformanceTestsProject.groovy
@@ -1,4 +1,6 @@
-package com.blackbaud.templates.tasks;
+package com.blackbaud.templates.project
+
+import com.blackbaud.templates.project.BasicProject;
 
 class PerformanceTestsProject {
