+++ src/main/groovy/com/blackbaud/templates/project/ProjectFile.groovy
@@ -1,4 +1,4 @@
-package com.blackbaud.templates
+package com.blackbaud.templates.project
 
 
 class ProjectFile extends File {
