+++ src/main/groovy/com/blackbaud/templates/project/RestProject.groovy
@@ -1,8 +1,6 @@
-package com.blackbaud.templates.tasks
+package com.blackbaud.templates.project
 
 import com.blackbaud.templates.CurrentVersions
-import com.blackbaud.templates.ProjectFile
-import com.blackbaud.templates.ProjectProps
 
 import static com.google.common.base.CaseFormat.LOWER_CAMEL
 import static com.google.common.base.CaseFormat.LOWER_HYPHEN
