+++ src/main/groovy/com/blackbaud/templates/tasks/AddJpaObjectTask.groovy
@@ -1,5 +1,6 @@
 package com.blackbaud.templates.tasks
 
+import com.blackbaud.templates.project.RestProject
 import org.gradle.api.tasks.TaskAction
